+++ attributes/freebsd.rb
@@ -8,7 +8,6 @@ if platform?('freebsd')
   default['filtermail']['bin'] = '/usr/local/bin/filtermail'
   default['chatmail']['certificates_dir'] = lazy { "#{node['lego']['path']}/certificates" }
   default['virtualenv'] = '/usr/local/bin/virtualenv'
-  default['chatmail']['python_version_string'] = '3.11'
   default['chatmail']['packages'] = %w( python3
                                         python
                                         devel/py-virtualenv
@@ -39,7 +38,9 @@ if platform?('freebsd')
   default['opendkim']['user'] = 'mailnull'
   default['opendkim']['group'] = 'mail'
   default['opendkim']['config_dir'] = lazy { "#{node['etcdir']}/mail" }
+  default['opendkim']['config_file'] = lazy { "#{node['opendkim']['config_dir']}/opendkim.conf" }
   default['opendkim']['pidfile'] = '/var/run/milteropendkim/pid'
+  default['opendkim']['genkey_bin'] = '/usr/local/sbin/opendkim-genkey'
   default['chatmail']['turnservice'] = 'chatmail_turn'
   default['unbound']['trust_anchor'] = lazy { "#{node['etcdir']}/unbound/root.key" }
   default['unbound']['anchor_bin'] = '/usr/local/sbin/unbound-anchor'