+++ recipes/sysctl.rb
@@ -6,6 +6,13 @@
 
 # Linux inotify settings
 if platform_family?('debian')
+  file '/etc/sysctl.conf' do
+    action :create_if_missing
+    owner 0
+    group 0
+    mode '0644'
+  end
+
   inotify_sysctls = %w(max_user_instances max_user_watches)
 
   inotify_sysctls.each do |s|