+++ CHANGELOG.md
@@ -7,6 +7,14 @@ This file is used to list changes made in each version of the chatmail cookbook.
 ### Changes
 
 - Dovecot: additional metadata reduction via aggressive header stripping before delivery
+- Chatmaild: Update to latest
+- OpenDKIM: Use absolute path to opendkim-genkey
+- Debian: Fix deployment path of OpenDKIM config file
+- Support Chef/CINC 19 (Ruby 3.2+)
+- Debian: Fix deployment of the Lego binary
+- Remove hardcoded Python version string
+- Debian: Custom install Dovecot packages like upstream (apt repo is deprecated)
+- Debian 13 (Trixie) support
 
 ## 0.3.1